Add shellcheck and a integration test
Couldn't get fedora/redhat type containers to work. TODO.

Shellcheck 2070 says:
Exceptions

If the strings happen to be version numbers and you're using <, or > to compare them as strings, and you consider this an acceptable thing to do, then you can ignore this warning.

So I left it. The alternatives are complicated bash, or external tools
that aren't always installed.